+++ bin/basel-crud.js
@@ -11,6 +11,8 @@ program
   .option('-db, --database <database>', 'Database')
   .option('-c, --controller <controller>', 'Controller name')
   .option('-v, --view <view>', 'View name (.html)')
+  .option('-r, --route <route>', 'Route (Ex.: persons)')
+  .option('-m, --menu <menu>', 'Show in main menu (1 - Yes, 0 - No)')
   .parse(process.argv);
 
 basel.config.name_ = program.args.length ? program.args[0] : 'basel';
@@ -18,6 +20,8 @@ basel.config.table = program.table;
 basel.config.database = program.database || basel.config.database;
 basel.config.controller = program.controller;
 basel.config.view = program.view;
+basel.config.route = program.route;
+basel.config.show_menu = program.menu || 1;
 
 
 if(basel.error){
@@ -40,6 +44,17 @@ function wizard (options) {
 		});
 	}
 
+	if(!options.route){
+
+		var word = options.name_.toLowerCase().replace(/ /g,'');
+
+		questions.push({
+			type: 'input',
+			name: 'route',
+			message: 'What is the route way?',
+			default: word
+		});
+	}
 
 	inquirer.prompt(questions, function(results) {
 	    _.assign(options, results);