+++ CHANGELOG.md
@@ -3,3 +3,5 @@ This change log is started in 0.0.9;
 
 ## 0.0.9 (2016-04-19)
 - Creating controller in basel-crud
+- Changing index in basel-crud
+- Review in table creating - basel-database